Adds contact form (uses Simple Form)
This is a lame approach to handling this, but given how rare emails are coming in anyway, this should work fine for the interim. It simply routes messages through [Simple Form](http://getsimpleform.com), which in turn delivers message to my inbox. Additionally, there is a dashboard view of recent messages for Administrators.