Remove / refactor class methods
Goal of this PR is to simplify the models back down to a reasonable size and rely more on the built-in capabilities of ActiveRecord relationships. It also cut down on duplicated code to do string formatting. The user invitation process is now under test, though there is some UI issue where the growl `flash` notifications are not displaying.