Update reply_to, from on mailers
Fixes #82

When using a third-party mailer, you won't want to change the raw "from" address because it will likely not be "approved" to send email on behalf of that domain. Instead, leave it as a `no-reply@example.com` and just update the `reply_to` attribute. This increases the deliverability of the email by reducing the chance it will be marked as spam.