Auto-populate event schedule on group creation
The SODA service works off a pre-paid budget of tokens for retrieving team schedules. To make it easier for new users, let's go ahead and retrieve the latest schedule available when they first create a group. Worst-case scenario is that the user is presented with the "Notify me when the schedule is available" screen instead of a list of events (which is the current behavior). We would then be able to handle any cases where the budget may have been depleted.