Add bulk price edit of tickets
Prior to this PR, a user that created season tickets with an incorrect price (or later decides to change the price) would have to click through to each ticket they held and update the pricing. This PR introduces a new `/tickets` route that allows the user to view all tickets at once for future events (can be toggled to past events) and update the pricing quickly in a spreadsheet-like format.

Fixes #111