Expand user profile
Fixes #45

Prior to this PR, a `User` object contained the first name, last name and email address of a user along with their preferred timzeone. This PR introduces three new fields via a one-to-one relationship with a new `Profile` object.

- bio (for open-ended details)
- location
- mobile

These can be expanded later for future stuff. I wanted to only add fields to the `User` object that directly affect the behavior of the user while in the system (e.g. timzone).