Add custom group avatar
Fixes #93

Prior to this PR, group users would not have any kind of visual indicator which group they were in in a rather text heavy interface. This PR adds the ability for a group administrator to set a custom avatar for the group. If one is not chosen, a default "ticket" image will be shown.

![screenshot](http://cl.ly/image/322q3V1A2E11/Screen%20Shot%202014-11-08%20at%2012.00.43%20AM.png)