Ticket request/assignment notifications via SMS
:exclamation::exclamation::exclamation:

Related to parts of #100

Prior to this PR, the only notification option offered for when a ticket was requested or assigned was via email. We need something a bit more robust than that, so I integrated with Twilio to send text messages. This PR introduce the `twilio-ruby` gem and associated configuration environment variables:

```bash
export TWILIO_ACCOUNT_SID=some_sid
export TWILIO_AUTH_TOKEN=some_token
export TWILIO_OUTBOUND_NUMBER=the_number
```

I also fixed a handful of minor annoyances with tests and the way the profile form behaved. I've run through the sandbox credentials and was able to get confirmation of a successful email send. Additionally, the `global_phone` gem was requred in order to meet Twilio's spec for all numbers to be sent in E.164 format (e.g. +15555551234) and to do that reliably for future phone numbers if we ever offer international messaging.

I've added HoundCI to this repository, so it will gripe about a bunch of stuff. I'll take care of the styleguide errors throughout the day tomorrow.