Add Rubocop to Guard installation
Adds the [guard-rubocop](https://github.com/yujinakayama/guard-rubocop) plugin so we can monitor the progress of syntax/style guide issues. Experience with it so far has been positive, even if I have NFI how to fix a good portion of the errors. I'll have a separate commit that brings down the count of issues from ~2300 to 237 by using the automatic syntax correction. All indications are it didn't break any of the tests, so that's :cool:. The configuration options for what to squelch/hide are
found in the `.rubocop.yml` file and in their documentation found [here](https://github.com/bbatsov/rubocop#configuration).