Massive ruby syntax cleanup
Side effect of installing Rubocop is that it gripes about a whole lot of stuff. So, I first ran the automatic code cleanup option `rubocop <target> -a` and then followed it up with plowing through each flagged line manually. The test suite remains in tact, so I'm relatively confident in how this turned out. Still, I'm not in a rush to merge this until I've had a better chance to do some UAT on the app locally to make sure the critical paths are working properly.

I stopped working on it when it had 99 problems detected, mostly because those are going to be harder to solve ... and because I enjoy a cheap laugh.

Details of remaining issues after this PR: https://gist.github.com/stephenyeargin/136403f7c0f4d50cfb42