Fix workflow for registering with a group code
Arafat registered recently using the link. Only trouble is that the link doesn't actually do anything special there, just makes a note of the invitation code. This fixes that by wiring in a URL that will properly handle the invitation code.

Four kinds of ways to register now:

- Generic home page registration - Takes you to the landing page to create or join a group
- Register with a team (entity) selected - Prefills the create group page
- Register with an individual group invitation - Prefills your unique invitation code
- Register with the group code - Prefills with the group's invitation code

This has room for cleanup in general, but those are the four major workflows.