Update group creation process
We have had some users join, but none have created a group. In looking as to why this is the case, I see that we drop them to a "Welcome" page that prompts them to either create or join a group. None of the sample size has clicked "Create a Group", and "Join a Group" is not useful unless you have an invitation code.

This PR guides them one step closer to create a group.

- New users are sent to the "Create a Group" page after account creation unless they used an invitation code
- The Group creation page receives some UI updates to guide them through the steps to creating a group
- The Welcome page is only shown if they elect not to create a group and attempt to go the primary route without having a group membership

\## Old

![](http://cl.ly/image/2N472e1a3x2g/Image%202015-03-01%20at%2012.08.16%20PM.png)

\## New

![](http://cl.ly/image/433g2E0E3c0Y/Screen%20Shot%202015-03-01%20at%2012.03.58%20PM.png)

This will break some of the Mixpanel goal tracking (it depended on viewing the Welcome page), but that is not relevant to our needs right now.