Display share-able group invitation code
Fixes #165 for @bval

- Adds link to the Invitation form
- Adds `zeroclipboard-rails` gem for usability++
- Fixes an issue where a user could potentially join a group as an admin and member (discovered in testing)

![](http://cl.ly/image/0Y472M0F422D/Image%202015-03-01%20at%208.37.50%20PM.png)

After working with the sign-up process, this made more sense to simply share as an alternative way to inviting a user. The email-based invites are not getting noticed (because email sucks). It makes more sense to do it this way. If we go another route with subscription limits, we would allow a Group Administrator to turn off invitations entirely (or when their subscription reached a particular limit). That's a problem for later.

The user-based invitation code creation remains on the table, but this removes the barrier for now.