Fix for broken update login/password
Fixes #172

This bug was introduced when attempting to bypass some of Devise's built-in sanity checks around registering for an account while you are already logged in. I wanted ot use a register route for the invitation codes that would transparently redirect to the appropriate place within the application in the event an existing, signed-in user clicked one of them. Trouble is that the `RegistrationController` is also used to handle login updates, which were not accounted for.

Updated Profile Flow test cases to cover this particular part of the app.