Add manual schedule upload from SODA
With the MLB schedules this year, the data provider was lousy about keeping/publishing the regular season schedules separate of the pre-season schedules. As such, we were going to have to wait until opening day to get the regular season schedules.

They resolved this by sending me a free ZIP file of all MLB schedules. As the current iteration of the importer was only written to communicate with the API, this PR adds the ability to import them as one-off files. Will be handy for development purposes later anyway.