Set default group icon based on entity type
The default ticket image is fine, but it's kind of irrelevant for a new user (and most are not changing it for their group). This PR adds a series of public domain clipart images for the various leagues to give the group a more personal feel. Users are still able to override the default image with their team or group logo.