Allow mass assignment of tickets
![](http://dropbox.stephenyeargin.com/Mk3If.png)

Prior to this PR, the only field that could be edited on the master Tickets view was the cost of the ticket. This PR introduces the ability to also edit the assigned user, meaning that a seat draft or other mass-updated process could be initiated from this page. I saw a note in one of the group invitations that it was their stated purpose, and this should make that process a bit easier to do rather than clicking through each ticket to assign it.

Required a bit of refactoring in the Tickets controller, which is getting way out of hand.