Fix broken User Alias delete / validation
While checking other things, I noticed that you could not delete a User Alias after creating it. This PR takes care of that by ensuring all the CRUD functionality is working as expected, and adds a note at ticket creation that User Aliases can be used to reserve tickets for non-users.