Replace SODA with SeatGeek as schedule provider
Fixes #206

It took quite a bit of experimenting, but I believe that we can make [SeatGeek](http://platform.seatgeek.com) work for all of our schedule retrieval needs. This PR does a few things of note:

- Removes SODA API retrieval
- Removes SODA file import (which is all I had been able to use)
- Removes attempt at auto-downloading a schedule on signup
- Adds the rather abandoned but functional SeatGeek gem
- Adds a Rake task suite to
 - Update entities import key that exactly match the available taxonomy
 - Update events where entities have an import key
- Adds `EntityType.import_key` via migration
- Adds import button in the admin per Entity (to do one-off data pulls)
- Adds a panel to the admin dashboard highlighting groups without events

There are still a handful of things that will need to happen before this will completely work.

- [ ] Add the import keys (which are URLs that get parsed into the gem) for the available entity types
- [ ] Rename (en masse) the teams for NCAAMB (append ` Basketball`) and NCAAWB (append `Womens Basketball`) so that the names match up.
- [ ] Make minor corrections (or manual edits) to team names (Anaheim Angels -> Los Angeles Angels, no accent marks in Montreal)

The schedule importer is set up to run nightly. Drawback is we will end up with a a lot more data that may never get used.