Remove Mixpanel Analytics
Removes the tracking code/integration for Mixpanel. The service itself is handy, but they are obviously aimed more at supporting a mobile application. Because of changes in #217, the user tracking hasn't worked properly since (creating several duplicate entries that can't be merged). That, and we were giving free advertising for them in order to get data points we weren't using for anything specific.

If we re-visit user analytics, I'd prefer it be something more tailor made for a SaaS product.