Allow CSV imports for events
Ran into a case where it was rather tedious to manually add events for a roller derby team. This PR allows for a rather basic import functionality, leveraging what is already in place for the SeatGeek importing.

Also fixes several CVEs that have come out in the last few weeks for Rails and Devise.