Unassign alias on bulk edit
Fixes #245

When a ticket had an alias set, it would not be cleared when an assignment was done through the bulk editor. This really should be handled through an `assign` method on the mode. In the mean time, this adds a line to the controller.

Also moved the `TicketHistory` operations into the model to reduce the size of that controller and get it under test.