Migrate int columns to boolean
When creating the original migration, I had used the integer type for `status` and other boolean columns. This PR converts those particular columns to their proper data type by way of migrating the data into a new column, dropping the old one, and renaming the new one to the name of the old one. I'm verifying this will the live database to check for errors that may not be accounted for in testing.