Add HeadwayApp Change Log Badge
Prior to this PR, there was not a good way to get updates about product changes while logged into the app. We usually rely on email campaigns or a user to stumble onto the blog. [HeadwayApp](https://headwayapp.co/) is a free service that publishes changelogs -- that's it. They are likely positioning themselves to get bought by another company, so there's no pricing plan for the product yet.

The changelog, in addition to being visible in the app, is available at http://changelog.myseatshare.com/