Allow a non-Grouped user to see own profile
If a user attempts to update their profile before joining a group, they are given an access denied message. The reason for this is that the profile page does a check before showing a particular profile that the logged in user must share a group with them. In the case of a user without a group, that check fails. This PR adds an exception that a user can always view their own profile and adds a relevant test case:

Fixes #286

Unrelated:

- Updated admin API to use better `assert_equal` instead of normal `assert`.